puppeteer: Fix `test_narrow_public_streams` test  in `message-basics.ts`.
It was introduced in
https://github.com/zulip/zulip/commit/6c1eb82e6f8bbb7c5b08b4d6883485d2b69d968e commit.

It fixes `test_narrow_public_streams` test which checks
that, there are no bookends in ` /streams/public`. The
issue was there were no bookends by default when we
were checking in the Denmark stream. We fix this by
unsubscribing from the Denmark stream which will
create a bookend.