templates: Improve text on 500 error pages.
Makes the 500 error page text conditional on whether the server is
self-hosted or Zulip Cloud.

In the case of Zulip Cloud, when `settings.CORPORATE_ENABLED` is true,
the text provides an email address to contact Zulip support.

In the case of self-hosted, when `settings.CORPORATE_ENABLED is false,
the text provides an email address to contact server administrators
and a link to Zulip's server troubleshooting guide.

Fixes #23063.