url_encoding: Use user name instead of email in personal_narrow_url.
This commit updates the urls for personal narrow sent in email
notifications to be of form "{user_id}-{encoded_full_name}" to
make it consistent with the urls that we use for such narrows
in webapp which were recently updated in 8b50328 for improving
performance. We encode the full name in the same way that we do in
webapp by replacing the url characters encoded by browser with "-".