compose: Increase z-index for compose box.
Following e7ab5ccede68d7de6799adcf906a96c37145a72a, the z-index of the
compose box (2) tied with that of unread date dividers within a topic;
this could lead to the dividers incorrectly appearing on top of the
compose box.

We don't have other elements that can overlap with compose with a
z-index between 2 and 4, so just increase it to 4.