compose_ui: Revert changes for replace_syntax done in b7c8128b7f.
This commit reverts the changes for replace_syntax in b7c8128b7f
and we again use JS replace instead of replace from text-field-edit.

We do this change because replace from text-field-edit leaves the
replaced text selected, which we don't want. This change is
temporary and we can probably use replace method from
text-field-edit once this issue is fixed in upstream.