util: Move is_pm_recipient to compose_fade_helper.
"is_pm_recipient" is not supposed to be called with an arbitrary
recipient object which might have "to_user_ids" being undefined.
Since this helper is only used with focused_recipient in
compose_fade_helper, we move it there.

Note that the helper is no longer separately tested. It is now covered
by the test case of "compose_fade_helper.would_receive_message".

(See
https://github.com/zulip/zulip/blob/201fdfc4b3700a52618f86c62c09b3c3e97e9aff/static/js/compose.js#L156-L162)

Signed-off-by: Zixuan James Li <p359101898@gmail.com>