settings: Fix bug when changing stream specific notification setting.
The class for "tr" element was changed in 2b29e43c06442d, so
get_stream_id function was not able to get the correct stream-id
from the target element. This commit fixes get_stream_edit
function to also check for closest ".stream-notifications-row"
ancestor element and thus it returns the correct stream id.