upload: Remove and reduce timeouts for closing upload bar.
This timeout was introduced in this commit: 913a09bed4c190d354faa644a17ae6477a581be5

The UI should close immediately when the user clicks cancel,
and the rest of the canceling code can run behind the scenes.

We want to keep a short timeout for upload completion
so that the user sees the 100% complete upload bar.