CVE-2023-22735: Set a Content-Security-Policy header on proxied S3 data.
This was missed in 8d8433d41341324f444667acf8124e0840c6b278; as this
content is fundamentally untrusted, it must be served with
`Content-Security-Policy` headers in order to be safe.  These headers
were not provided previously for S3 content because it was served from
the S3 domain.

This mitigates content served from Zulip which could be a stored XSS,
but only in browsers which support Content-Security-Policy headers;
see subsequent commit for the complete solution.