stream_popover: Remove useless old_topic_name check.
This has been present since this modal was first introduced in
0a6d1de69ee8d7a0131da79d324c12abeefd149e, but as far as I can tell, it
has never been correct. We know `old_topic_name` is not
null/undefined, since we do a check with it trimmed earlier in the
function, and there is no product reason why we would would to
silently fail to move a topic because its name was the empty string.