tooltips: Fix hotkey-hint CSS class reuse.
The "hotkey-hint" class name used internally in
89d2659d369ae99b0e2e6a98e054f99ee74de7a9 conflicted with the
preexisting class name for hotkey hints into popovers, introduced in
f317fbabce153d849a21a7fa5318afcc43633f27.

Given that the new class is for a styling of hotkey hints designed for
use in tooltips, it was a bad name anyway, so just rename it to
tooltip-hotkey-hint. We rename the related cluster of variable names
to match this.