data_import: Stop tar'ing up converted data.
`./manage.py import` does not take a tarball; it takes a directory.
Making a separate tarball is a waste of CPU time and disk, as it is
never used.

This was included in the commit of the initial Slack conversion code
in 283d46c9bfdc and propagated from there into every conversion tool.

Remove the unnecessary tarball creation.